requirements.txt
audioread==2.1.9
audiosegment==0.23.0
docker==4.1.0
docker-compose==1.25.0
ffmpeg==1.4
keras==2.9.0
librosa==0.9.1
matplotlib==3.5.3
numpy==1.22.4
opencv-python==4.6.0.66
pandas==1.4.2
psycopg2-binary==2.9.3
pygame==2.1.2
scikit-learn==1.1.2
tensorflow==2.9.1
pymongo==4.3.3